+++ app/src/main/java/by/geekbrains/appnotes/domain/NoteRepository.java
@@ -1,0 +1,10 @@
+package by.geekbrains.appnotes.domain;
+
+import java.util.ArrayList;
+
+public interface NoteRepository {
+
+    ArrayList<NoteEntity> getNotes();
+
+    void deleteNote(NoteEntity noteEntity);
+}
